SW470068: Add support for enabling/disabling network IPMI
The changes have been pulled from OP930 which was dropped during refactoring
of the IPMI user/channel management. Updating the channel privilege on the
network interface is removed since the change is not available in the upstream
and is not mandatory for this feature. The following issue is tracking this.

https://github.com/openbmc/phosphor-host-ipmid/issues/139#issuecomment-533550121

A solution on OpenBMC master would be to upstream the service configuration
manager from Intel to handle enabling/disabling of different services.

https://github.com/Intel-BMC/provingground/tree/master/srvcfg-manager

Change-Id: Iec18d9a9d75033a43474506a2c9f2252ad87cd26
Signed-off-by: Tom Joseph <tomjoseph@in.ibm.com>