SW460600: Don't report serviceable action if the watchdog service is not present
There were a few cases where a service action was reported because the
watchdog service was not present. It was because of a race condition where
the power down operation is trigerred and the watchdog service is shutdown
but if was followed by a watchdog reset. Since the watchdog service is absent
it resulted in an error log which calls for serviceable action. The IPMI error
response code will report the failure to the host platform software.

Change-Id: Iae1026f3f7b56c422ad35bec50db746eda07171d
Signed-off-by: Tom Joseph <tomjoseph@in.ibm.com>