Add new oemibmserviceagent privilege role
This enhances phosphor-host-ipmi to accept the new Phosphor privilege role
oemibmserviceagent which corresponds to the OemIBMServiceAgent Redfish role.

Tested:
Minimal testing.  Without this, if any local BMC user has the
priv-oemibmserviceagent role, this service would fail with "Error in
converting to IPMI privilege".
With this fix, the user privilege role is accepted.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>