user_mgmt: set priv-user as the default privilege
Due wrong interpretation of the IPMI spec there is appeared the
`no-access` privilege in the OpenBMC. But it's a state of unconfigured
channel and shouldn't be used as the valid user's privilege.

This commit is a part of the patch set removing `priv-noaccess` from
OpenBMC. This replaces the default privilege for newly created users
from `priv-noaccess` to `priv-user`.

This doesn't affect for users with already privilege set, but it's
highly recommended to change their role from `priv-noaccess` to some
valid privilege or delete them at all.

Tested: New users create by IPMI command have `USER` privilege by
        default.
        Command: ipmitool user set name <user id> <username>
        Response: // User created successfully
        Command: ipmitool user list
        Response: // The channel privilege level for this new user
                     contains `USER`.

Change-Id: I4b99551448245d3a9ed0882f581784a0ee35e3d8
Signed-off-by: Alexander Filippov <a.filippov@yadro.com>