dcmi: fix Get Sensor Info command
Issue: if users request to Get Sensors Info via "ipmitool dcmi sensors"
command, the Record ID of sensors always are 0x0000.

Root Cause: In the loop of scanning the configuration file to get the
record ID, it is using the "config" variable instead of the "reading"
variable. It makes the record ID always is 0x0000.

Solution: Replace the "config" variable with the "reading" variable.

Tested:
    1. Request to Get Sensors Info.
       $ipmitool dcmi sensors
    2. The list of sensors and temperature are shown.

Change-Id: If2b03b6b8a077a4e9239403ec9ecd640f645cbc0
Signed-off-by: Thang Tran <thuutran@amperecomputing.com>